--- src/renderer/src/dashboard.tsx
@@ -1,12 +1,0 @@
-import Versions from './components/Versions'
-
-function Dashboard(): JSX.Element {
-  return (
-    <div className="container">
-      <Versions></Versions>
-      HELLO DASHBOARD
-    </div>
-  )
-}
-
-export default Dashboard
